+++ tests/hosted/src/margin.rs
@@ -25,6 +25,7 @@ use anyhow::{bail, Error};
 
 use jet_margin::{MarginAccount, PositionKind};
 use jet_margin_sdk::ix_builder::{ControlIxBuilder, MarginPoolConfiguration, MarginPoolIxBuilder};
+use jet_margin_sdk::swap::SwapPool;
 use solana_sdk::instruction::Instruction;
 use solana_sdk::signature::{Keypair, Signer};
 use solana_sdk::system_program;
@@ -36,7 +37,7 @@ use jet_margin_sdk::tx_builder::MarginTxBuilder;
 use jet_metadata::{LiquidatorMetadata, MarginAdapterMetadata, TokenKind, TokenMetadata};
 use jet_simulation::{send_and_confirm, solana_rpc_api::SolanaRpcClient};
 
-use crate::{swap::SwapPool, tokens::TokenOracle};
+use crate::tokens::TokenOracle;
 
 /// Information needed to create a new margin pool
 pub struct MarginPoolSetupInfo {
