+++ tests/hosted/tests/swap.rs
@@ -1,5 +1,6 @@
 use anyhow::Error;
 
+use jet_margin_sdk::swap::SwapPool;
 use jet_static_program_registry::{orca_swap_v1, orca_swap_v2, spl_token_swap_v2};
 use solana_sdk::native_token::LAMPORTS_PER_SOL;
 use solana_sdk::pubkey::Pubkey;
@@ -8,7 +9,7 @@ use solana_sdk::signature::Signer;
 use hosted_tests::{
     context::{test_context, MarginTestContext},
     margin::MarginPoolSetupInfo,
-    swap::SwapPool,
+    swap::SwapPoolConfig,
     tokens::TokenPrice,
 };
